refactor(web): 웹 소스를 src 아래로 옮기고 기능별 폴더 구조로 정리
