--- apps/web/shared/components/layouts/index.ts
@@ -1,1 +1,0 @@
-export {};
