--- apps/web/shared/index.ts
@@ -1,3 +1,0 @@
-export { LanguageSwitcher } from "./components/language-switcher";
-export * from "./components/layouts";
-export { Providers } from "./components/providers";
