+++ apps/web/src/app/[locale]/(auth)/register/page.tsx
@@ -17,9 +17,7 @@ export default async function RegisterPage({ params }: RegisterPageProps) {
 				<div className="text-center">
 					<h1 className="text-2xl font-bold">{t("register")}</h1>
 				</div>
-				<p className="text-center text-gray-500">
-					회원가입 폼이 여기에 표시됩니다.
-				</p>
+				<p className="text-center text-gray-500">회원가입 폼이 여기에 표시됩니다.</p>
 				<p className="text-center text-sm text-gray-600">
 					{t("hasAccount")}{" "}
 					<Link href="/login" className="text-black underline">
