+++ apps/web/src/i18n/request.ts
@@ -4,15 +4,12 @@ import { routing } from "./routing";
 export default getRequestConfig(async ({ requestLocale }) => {
 	let locale = await requestLocale;
 
-	if (
-		!locale ||
-		!routing.locales.includes(locale as (typeof routing.locales)[number])
-	) {
+	if (!locale || !routing.locales.includes(locale as (typeof routing.locales)[number])) {
 		locale = routing.defaultLocale;
 	}
 
 	return {
 		locale,
-		messages: (await import(`../messages/${locale}.json`)).default,
+		messages: (await import(`./messages/${locale}.json`)).default,
 	};
 });
