+++ apps/web/src/i18n/routing.ts
@@ -8,5 +8,4 @@ export const routing = defineRouting({
 	localePrefix: "as-needed",
 });
 
-export const { Link, redirect, usePathname, useRouter, getPathname } =
-	createNavigation(routing);
+export const { Link, redirect, usePathname, useRouter, getPathname } = createNavigation(routing);
